ParallelProject_152627/src/main/java/com/capg/bean/AccountDetails.java
package com.capg.bean;

import java.util.List;

public class AccountDetails {

	private long accNo;
	private String username;
	private String password;
	private double balance;
	private List<String> transId;
	private String accType;
	
	private CustomerDetails customer;
	
	public long getAccNo() {
		return accNo;
	}
	public void setAccNo(long accNo) {
		this.accNo = accNo;
	}

	
	public String getUsername() {
		return username;
	}
	public void setUsername(String username) {
		this.username = username;
	}
	public String getPassword() {
		return password;
	}
	public void setPassword(String password) {
		this.password = password;
	}
	
	
	public double getBalance() {
		return balance;
	}

	public void setBalance(double balance) {
		this.balance = balance;
	}
	
	public CustomerDetails getCustomer() {
		return customer;
	}
	public void setCustomer(CustomerDetails customer) {
		this.customer = customer;
	}
	public List<String> getTransId() {
		return transId;
	}
	public void setTransId(List<String> transId) {
		this.transId = transId;
	}

	public String getAccType() {
		return accType;
	}
	public void setAccType(String accType) {
		this.accType = accType;
	}
	
	

	
}

